+++ src/main/java/it/uniroma2/dicii/ispw/controller/GestioneUtentiController.java
@@ -2,13 +2,12 @@ package it.uniroma2.dicii.ispw.controller;
 
 
 import it.uniroma2.dicii.ispw.App;
-import it.uniroma2.dicii.ispw.bean.AnnouncementBean;
 import it.uniroma2.dicii.ispw.bean.CorsoBean;
 import it.uniroma2.dicii.ispw.bean.UtenteBean;
 import it.uniroma2.dicii.ispw.enums.TypesOfPersistenceLayer;
 import it.uniroma2.dicii.ispw.exception.InvalidDataException;
+import it.uniroma2.dicii.ispw.exception.ItemAlreadyExistsException;
 import it.uniroma2.dicii.ispw.exception.ItemNotFoundException;
-import it.uniroma2.dicii.ispw.model.announcement.Announcement;
 import it.uniroma2.dicii.ispw.model.corso.Corso;
 import it.uniroma2.dicii.ispw.model.corso.dao.CorsoDAO;
 import it.uniroma2.dicii.ispw.model.corso.dao.CorsoDBMS;
@@ -18,22 +17,13 @@ import it.uniroma2.dicii.ispw.model.utente.Utente;
 import it.uniroma2.dicii.ispw.model.utente.dao.UtenteDAO;
 import it.uniroma2.dicii.ispw.model.utente.dao.UtenteDBMS;
 import it.uniroma2.dicii.ispw.model.utente.dao.UtenteFS;
-import it.uniroma2.dicii.ispw.utils.LoggerManager;
+import it.uniroma2.dicii.ispw.utils.EmailValidator;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 public class GestioneUtentiController {
     private static final int CF_LENGTH = 16;
-    private static final String EMAIL_PATTERN =
-            "^[A-Z0-9._%+-]+@[A-Z0-9.-]+\\.[A-Z]{2,6}$";
-    private static final Pattern pattern = Pattern.compile(EMAIL_PATTERN, Pattern.CASE_INSENSITIVE);
-
     private UtenteDAO utenteDAO;
     private CorsoDAO corsoDAO;
     private UtenteCorsoDAO utenteCorsoDAO;
@@ -48,40 +38,21 @@ public class GestioneUtentiController {
         }
     }
 
-    public static boolean validate(String email) {
-        Matcher matcher = pattern.matcher(email);
-        return matcher.matches();
-    }
-
-    public static Date parseStringToDate(String str) throws InvalidDataException {
-        SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy");
-        Date date = null;
-        try {
-            date  = sdf.parse(str);
-        } catch (ParseException e) {
-            throw new InvalidDataException("Formato data non valida" + e.getMessage());
-        }
-        return date;
-    }
-
     public String insertUtente(UtenteBean utenteBean) throws InvalidDataException{
         String res = null;
-        if(utenteBean.getCf().length() != CF_LENGTH) throw new InvalidDataException("Il codice fiscale non è valido!");
+        if(utenteBean.getCf().length() != CF_LENGTH) throw new InvalidDataException("Codice fiscale non valido.");
         if(utenteBean.getName().isBlank() || utenteBean.getSurname().isBlank())
-            throw new InvalidDataException("Nome e/o cognome mancanti");
-        if(utenteBean.getEmail() == null || !validate(utenteBean.getEmail()))
-            throw new InvalidDataException("Email non valida");
+            throw new InvalidDataException("Nome e/o Cognome mancanti.");
+        if(utenteBean.getEmail() == null || EmailValidator.isNotValid(utenteBean.getEmail()))
+            throw new InvalidDataException("Email non valida.");
         if(utenteBean.getPassword() == null)
-            throw new InvalidDataException("La password non può essere vuota");
+            throw new InvalidDataException("Password mancante.");
 
-        SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy");
-        Date date = parseStringToDate(utenteBean.getBirthDate());
-        Utente utente = new Utente(utenteBean.getName(), utenteBean.getSurname(), utenteBean.getCf(), date, utenteBean.getEmail(), utenteBean.getPassword(), utenteBean.getRuolo());
+        Utente utente = new Utente(utenteBean);
         try {
             res = utenteDAO.insertUtente(utente);
-        } catch (Exception e) {
-            //error connection with persistence layer
-            throw new RuntimeException(e);
+        } catch (ItemAlreadyExistsException e) {
+            res = e.getMessage();
         }
         return res;
     }
@@ -90,43 +61,33 @@ public class GestioneUtentiController {
         List<Utente> users;
         List<UtenteBean> utenteBeans = new ArrayList<>();
 
-        try {
-            users = utenteDAO.getAllUtenti();
-        } catch (Exception e) {
-            LoggerManager.logSevereException(e.getMessage(), e);
-            return utenteBeans;
-        }
+        users = utenteDAO.getAllUtenti();
 
         for (Utente u: users) {
             UtenteBean ub = new UtenteBean(u.getName(), u.getSurname(), u.getCf(), u.getBirthDate().toString(), u.getEmail(), null, u.getRuolo());
             utenteBeans.add(ub);
         }
+
         return utenteBeans;
     }
 
-    public String editUtente(UtenteBean utenteBean) throws Exception {
-        String res = "";
-        Utente utente = utenteDAO.getUtenteById(utenteBean.getCf());
-        if (utente == null) throw new ItemNotFoundException("Utente sconosciuto");
+    public String updateUtente(UtenteBean utenteBean) throws InvalidDataException, ItemNotFoundException {
+        Utente utente = utenteDAO.getUtenteByCf(utenteBean.getCf());
+        if (utente == null) throw new ItemNotFoundException("Utente non trovato.");
+
         if(utenteBean.getName().isBlank() || utenteBean.getSurname().isBlank())
-            throw new InvalidDataException("Nome e/o cognome mancanti");
-        if(utenteBean.getEmail() == null || !validate(utenteBean.getEmail()))
-            throw new InvalidDataException("Email non valida");
+            throw new InvalidDataException("Nome e/o Cognome mancanti.");
+        if(utenteBean.getEmail() == null || EmailValidator.isNotValid(utenteBean.getEmail()))
+            throw new InvalidDataException("Email non valida.");
 
-        utente.setName(utenteBean.getName());
-        utente.setSurname(utenteBean.getSurname());
-        utente.setEmail(utenteBean.getEmail());
-        utente.setBirthDate(parseStringToDate(utenteBean.getBirthDate()));
-        utente.setRuolo(utenteBean.getRuolo());
-
-        res = utenteDAO.editUtente(utente);
-
-        return res;
+        utente.updateInfos(utenteBean);
+        return utenteDAO.editUtente(utente);
     }
 
-    public List<CorsoBean> getEnrollmentsByUtente(UtenteBean utenteBean) throws Exception {
-        Utente utente = utenteDAO.getUtenteById(utenteBean.getCf());
+    public List<CorsoBean> getEnrollmentsByUtente(UtenteBean utenteBean) throws ItemNotFoundException {
         List<CorsoBean> corsoBeanList = new ArrayList<>();
+
+        Utente utente = utenteDAO.getUtenteByCf(utenteBean.getCf());
         for (Corso c: utente.getEnrollments()) {
             CorsoBean corsoBean = new CorsoBean(c.getName());
             corsoBeanList.add(corsoBean);
@@ -134,24 +95,31 @@ public class GestioneUtentiController {
         return corsoBeanList;
     }
 
-    public String removeEnrollmentByUtente(UtenteBean utenteBean, CorsoBean corsoBean) throws Exception {
-        Utente utente = utenteDAO.getUtenteById(utenteBean.getCf());
+    public void removeEnrollmentByUtente(UtenteBean utenteBean, CorsoBean corsoBean) throws ItemNotFoundException {
+        Utente utente = utenteDAO.getUtenteByCf(utenteBean.getCf());
         Corso corso = corsoDAO.getCorsoByNome(corsoBean.getName());
-        return utenteCorsoDAO.removeEnrollmentByUtente(utente, corso);
+        if(utente != null && corso != null) {
+            utenteCorsoDAO.removeEnrollmentByUtente(utente, corso);
+        }
     }
 
-    public String addEnrollmentToUtente(UtenteBean utenteBean, CorsoBean corsoBean) throws Exception {
-        Utente utente = utenteDAO.getUtenteById(utenteBean.getCf());
+    public void addEnrollmentToUtente(UtenteBean utenteBean, CorsoBean corsoBean) throws ItemNotFoundException {
+        Utente utente = utenteDAO.getUtenteByCf(utenteBean.getCf());
         Corso corso = corsoDAO.getCorsoByNome(corsoBean.getName());
-        return utenteCorsoDAO.addEnrollmentToUtente(utente, corso);
+        if(utente != null && corso != null) {
+            utenteCorsoDAO.addEnrollmentToUtente(utente, corso);
+        }
     }
 
-    public List<CorsoBean> getTeachingByUtente(UtenteBean utenteBean) throws Exception {
-        Utente utente = utenteDAO.getUtenteById(utenteBean.getCf());
+    public List<CorsoBean> getTeachingByUtente(UtenteBean utenteBean) throws ItemNotFoundException {
         List<CorsoBean> corsoBeanList = new ArrayList<>();
-        for (Corso c: utente.getTeachings()) {
-            CorsoBean corsoBean = new CorsoBean(c.getName());
-            corsoBeanList.add(corsoBean);
+
+        Utente utente = utenteDAO.getUtenteByCf(utenteBean.getCf());
+        if(utente != null) {
+            for (Corso c : utente.getTeachings()) {
+                CorsoBean corsoBean = new CorsoBean(c.getName());
+                corsoBeanList.add(corsoBean);
+            }
         }
         return corsoBeanList;
     }
